+++ sdk/go/common/resource/plugin/analyzer_plugin.go
@@ -41,10 +41,11 @@ import (
 
 // analyzer reflects an analyzer plugin, loaded dynamically for a single suite of checks.
 type analyzer struct {
-	ctx    *Context
-	name   tokens.QName
-	plug   *plugin
-	client pulumirpc.AnalyzerClient
+	ctx     *Context
+	name    tokens.QName
+	plug    *plugin
+	client  pulumirpc.AnalyzerClient
+	version string
 }
 
 var _ Analyzer = (*analyzer)(nil)
@@ -83,7 +84,8 @@ func NewAnalyzer(host Host, ctx *Context, name tokens.QName) (Analyzer, error) {
 func NewPolicyAnalyzer(
 	host Host, ctx *Context, name tokens.QName, policyPackPath string, opts *PolicyAnalyzerOptions) (Analyzer, error) {
 
-	proj, err := workspace.LoadPolicyPack(filepath.Join(policyPackPath, "PulumiPolicy.yaml"))
+	projPath := filepath.Join(policyPackPath, "PulumiPolicy.yaml")
+	proj, err := workspace.LoadPolicyPack(projPath)
 	if err != nil {
 		return nil, errors.Wrapf(err, "failed to load Pulumi policy project located at %q", policyPackPath)
 	}
@@ -119,8 +121,15 @@ func NewPolicyAnalyzer(
 	// program directory, so that the '@pulumi/pulumi/cmd/run-policy-pack' module from the policy pack's
 	// node_modules is used.
 	pwd := policyPackPath
-	plug, err := newPlugin(ctx, pwd, pluginPath, fmt.Sprintf("%v (analyzer)", name),
-		[]string{host.ServerAddr(), "."}, env)
+
+	args := []string{host.ServerAddr(), "."}
+	for k, v := range proj.Runtime.Options() {
+		if vstr := fmt.Sprintf("%v", v); vstr != "" {
+			args = append(args, fmt.Sprintf("-%s=%s", k, vstr))
+		}
+	}
+
+	plug, err := newPlugin(ctx, pwd, pluginPath, fmt.Sprintf("%v (analyzer)", name), args, env)
 	if err != nil {
 		// The original error might have been wrapped before being returned from newPlugin. So we look for
 		// the root cause of the error. This won't work if we switch to Go 1.13's new approach to wrapping.
@@ -136,10 +145,11 @@ func NewPolicyAnalyzer(
 	contract.Assertf(plug != nil, "unexpected nil analyzer plugin for %s", name)
 
 	return &analyzer{
-		ctx:    ctx,
-		name:   name,
-		plug:   plug,
-		client: pulumirpc.NewAnalyzerClient(plug.Conn),
+		ctx:     ctx,
+		name:    name,
+		plug:    plug,
+		client:  pulumirpc.NewAnalyzerClient(plug.Conn),
+		version: proj.Version,
 	}, nil
 }
 
@@ -184,7 +194,7 @@ func (a *analyzer) Analyze(r AnalyzerResource) ([]AnalyzeDiagnostic, error) {
 	failures := resp.GetDiagnostics()
 	logging.V(7).Infof("%s success: failures=#%d", label, len(failures))
 
-	diags, err := convertDiagnostics(failures)
+	diags, err := convertDiagnostics(failures, a.version)
 	if err != nil {
 		return nil, errors.Wrap(err, "converting analysis results")
 	}
@@ -257,7 +267,7 @@ func (a *analyzer) AnalyzeStack(resources []AnalyzerStackResource) ([]AnalyzeDia
 	failures := resp.GetDiagnostics()
 	logging.V(7).Infof("%s.AnalyzeStack(...) success: failures=#%d", a.label(), len(failures))
 
-	diags, err := convertDiagnostics(failures)
+	diags, err := convertDiagnostics(failures, a.version)
 	if err != nil {
 		return nil, errors.Wrap(err, "converting analysis results")
 	}
@@ -325,10 +335,17 @@ func (a *analyzer) GetAnalyzerInfo() (AnalyzerInfo, error) {
 		}
 	}
 
+	// The version from PulumiPolicy.yaml is used, if set, over the version from the response.
+	version := resp.GetVersion()
+	if a.version != "" {
+		version = a.version
+		logging.V(7).Infof("Using version %q from PulumiPolicy.yaml", version)
+	}
+
 	return AnalyzerInfo{
 		Name:           resp.GetName(),
 		DisplayName:    resp.GetDisplayName(),
-		Version:        resp.GetVersion(),
+		Version:        version,
 		SupportsConfig: resp.GetSupportsConfig(),
 		Policies:       policies,
 		InitialConfig:  initialConfig,
@@ -596,10 +613,16 @@ func convertConfigSchema(schema *pulumirpc.PolicyConfigSchema) *AnalyzerPolicyCo
 	}
 }
 
-func convertDiagnostics(protoDiagnostics []*pulumirpc.AnalyzeDiagnostic) ([]AnalyzeDiagnostic, error) {
+func convertDiagnostics(protoDiagnostics []*pulumirpc.AnalyzeDiagnostic, version string) ([]AnalyzeDiagnostic, error) {
 	diagnostics := make([]AnalyzeDiagnostic, len(protoDiagnostics))
 	for idx := range protoDiagnostics {
 		protoD := protoDiagnostics[idx]
+
+		// The version from PulumiPolicy.yaml is used, if set, over the version from the diagnostic.
+		policyPackVersion := protoD.PolicyPackVersion
+		if version != "" {
+			policyPackVersion = version
+		}
 
 		enforcementLevel, err := convertEnforcementLevel(protoD.EnforcementLevel)
 		if err != nil {
@@ -609,7 +632,7 @@ func convertDiagnostics(protoDiagnostics []*pulumirpc.AnalyzeDiagnostic) ([]Anal
 		diagnostics[idx] = AnalyzeDiagnostic{
 			PolicyName:        protoD.PolicyName,
 			PolicyPackName:    protoD.PolicyPackName,
-			PolicyPackVersion: protoD.PolicyPackVersion,
+			PolicyPackVersion: policyPackVersion,
 			Description:       protoD.Description,
 			Message:           protoD.Message,
 			Tags:              protoD.Tags,